refactor(metadata): replace IdSharpMetaTag with NativeId3MetaTag and remove IdSharp dependency
- Migrated from `IdSharpMetaTag` to `NativeId3MetaTag` for improved .mp3 metadata handling.
- Removed `IdSharp` dependencies (`IdSharp.Common` and `IdSharp.Tagging`) across the project.
- Updated `DirectoryProcessorToStagingService` to use `NativeId3MetaTag`.
- Adjusted tests to reflect the metadata plugin migration.
- Cleaned up unused `AuralizeBlazor` dependency in the Blazor project.
