+++ Directory.Packages.props
@@ -7,7 +7,6 @@
     <PackageVersion Include="Asp.Versioning.Http" Version="10.0.0" />
     <PackageVersion Include="Asp.Versioning.Mvc" Version="10.0.0" />
     <PackageVersion Include="Asp.Versioning.Mvc.ApiExplorer" Version="10.0.0" />
-    <PackageVersion Include="AuralizeBlazor" Version="3.0.4" />
     <PackageVersion Include="BCrypt.Net-Next" Version="4.2.0" />
     <PackageVersion Include="Blazored.SessionStorage" Version="2.4.0" />
     <PackageVersion Include="bunit" Version="2.7.2" />
@@ -31,14 +30,13 @@
     <PackageVersion Include="HtmlAgilityPack" Version="1.12.4" />
     <PackageVersion Include="HtmlSanitizer" Version="9.1.923-beta" />
     <PackageVersion Include="IdGen" Version="3.0.7" />
-    <PackageVersion Include="IdSharp.Common" Version="2.0.3" />
-    <PackageVersion Include="IdSharp.Tagging" Version="2.0.3" />
     <PackageVersion Include="JetBrains.Annotations" Version="2025.2.4" />
     <PackageVersion Include="Jint" Version="4.10.0" />
     <PackageVersion Include="K4os.Hash.xxHash" Version="1.0.8" />
     <PackageVersion Include="LiteDB" Version="5.0.21" />
     <PackageVersion Include="Mapster" Version="10.0.8" />
     <PackageVersion Include="Markdig" Version="1.3.0" />
+    <PackageVersion Include="MessagePack" Version="3.1.7" />
     <PackageVersion Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="10.0.9" />
     <PackageVersion Include="Microsoft.AspNetCore.Components.Authorization" Version="10.0.9" />
     <PackageVersion Include="Microsoft.AspNetCore.Http" Version="2.3.11" />
