+++ benchmarks/Melodee.Benchmarks/Melodee.Benchmarks.csproj
@@ -17,6 +17,9 @@
     <PackageReference Include="BenchmarkDotNet" />
     <PackageReference Include="DecentDB.AdoNet" />
     <PackageReference Include="NBomber" />
+    <PackageReference Include="MessagePack">
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" />
     <PackageReference Include="FluentAssertions" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
