+++ docs/pages/changelog.md
@@ -40,6 +40,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MusicBrainz query probes now keep the broad ordered first-row existence
   measurement opt-in via `--include-row-existence` so default validation stays
   focused on request-safe indexed lookups.
+- Replaced the legacy IdSharp metadata fallback with Melodee's native ID3 tag
+  reader, removing the obsolete transitive image-processing dependency path.
+- Added explicit private `MessagePack` references for NBomber consumers so the
+  test and benchmark graphs resolve the non-vulnerable package version.
 
 ### Fixed
 
@@ -139,7 +143,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **Replaced `SixLabors.ImageSharp` with `SkiaSharp`** for all image processing operations.
+- **Replaced the legacy image processing library with `SkiaSharp`** for all image processing operations.
   A new `IImageProcessor` abstraction centralizes decode, encode, resize, format
   identification, and average-hash computation. `ImageHasher`, `ImageConvertor`, and
   `ImageValidator` now receive `IImageProcessor` via dependency injection rather than
